abstract out cache a bit more; add local cache into the mix
