+++ internal/cache/cache.go
@@ -1,11 +1,7 @@
 package cache
 
 import (
-	"context"
-	"fmt"
 	"net/url"
-
-	"github.com/go-redis/redis/v8"
 )
 
 type Cache interface {
@@ -17,69 +13,13 @@ type Cache interface {
 	CleanAll()
 }
 
-type RedisCache struct {
-	conn *redis.Client
-	ctx  context.Context
-}
-
-var cfg struct {
-	ConnectionString string
-}
-
 func GetCachingMechanism(connStr string) Cache {
 	uri, _ := url.Parse(connStr)
-	password, _ := uri.User.Password()
 
-	cch := &RedisCache{
-		conn: redis.NewClient(&redis.Options{
-			Addr:     fmt.Sprintf("%s:%s", uri.Hostname(), uri.Port()),
-			Username: uri.User.Username(),
-			Password: password,
-		}),
+	switch uri.Scheme {
+	case "redis":
+		return GetRedisCache(connStr)
+	default:
+		return GetLocalCache()
 	}
-
-	cch.ctx = context.Background()
-
-	return cch
-}
-
-func (rc *RedisCache) Put(key string, value interface{}) {
-	if err := rc.conn.Set(rc.ctx, key, value, 0); err != nil {
-		fmt.Println(err)
-	}
-}
-
-func (rc *RedisCache) PutAll(entries map[string]interface{}) {
-	for k, v := range entries {
-		rc.Put(k, v)
-	}
-}
-
-func (rc *RedisCache) Get(key string) interface{} {
-	value, err := rc.conn.Get(rc.ctx, key).Result()
-	if err != nil {
-		fmt.Println(err)
-		return ""
-	}
-	return value
-}
-
-func (rc *RedisCache) GetAll(keys []string) map[string]interface{} {
-	entries := make(map[string]interface{})
-	for _, k := range keys {
-		entries[k] = rc.Get(k)
-	}
-
-	return entries
-}
-
-func (rc *RedisCache) Clean(key string) {
-	if err := rc.conn.Del(rc.ctx, key); err != nil {
-		fmt.Println(err)
-	}
-}
-
-// CleanAll cleans the entire cache.
-func (rc *RedisCache) CleanAll() {
-	rc.conn.FlushDB(rc.ctx)
 }
